Add argument Bruker support indicated was necessary
From support:

The -AddRawFile or -arf command is broken/unusable, the logic is wrong so that it fails if the filelist file has a .txt extension, which it always will.

Instead use the -isf command followed by the -arfws command and pass the data set path without the filelist filename.  It does the same thing and doesn't fail.
I'll go ahead and fix the broken command for a future release.